Small improvement; js_only.
